==> 5_Predictor.ipynb <==
# Park Run  target setter

Import functions and required libraries

from functions.predictor import process_parkrun_data_for_models, user_input, fetch_runner_data, target_time

Update the filepaths below if you want to use different data/models:

model_to_use = 'models/to_use/xgb_opt_model.pkl'
scaler_to_use = 'models/to_use/minmax_scaler.pkl'
data_for_model = 'data/clean/cleaned_parkrun_no_outliers.csv'

df = process_parkrun_data_for_models(data_for_model)

## Automatic input (with parkrun id)

If you know your parkrun id, input it below for easier prediction. 

Should be an integer following A. For example A-5125087

Your link for full results should be in the form:
https://www.parkrun.org.uk/parkrunner/5125087/all

# Required parameter
parkrun_id = 5125087

Below are the optional parameters.

- next_parkrun_date: The date of the upcoming event in the form "YYYY-MM-DD"

Weather parameters for the anticipated conditions for the run in the form:
- temp: Temperature in °C
- wind: Windspeed in km/h
- precipitation: The expected precipitation amount in mm

If these are left blank the model will default to today's date and average values for the weather parameters.

# Optional parameters:
next_parkrun_date = '2024-12-21' # If not used then uses today's date

# Weather parameters. If not supplied then uses default/average values 
temp = 10
wind = 25
precipitation = 0

weather_list = [temp, wind, precipitation]

Get runner data and put through model for the run target:

runner_df = fetch_runner_data(parkrun_id, df, next_parkrun_date, weather_list)

target_time(runner_df, df, model_to_use, scaler_to_use)

## Manual input:

If you don't have your parkrun id, use the below function which will prompt you to enter the features required for a run time prediction.

user_stats = user_input(df)

target_time(user_stats, df, model_to_use, scaler_to_use)
